# Optimize aarch64 rotate